new really version  (3.1.5~39~really~3.1.4-20-gec8692283-1
* New snapshot: v3.1.4-20-gec8692283 works
* last working calamares snapshot, 21 introduce segfaults

+++ src/modules/users/SetPasswordJob.h
@@ -33,7 +33,7 @@ public:
     QString prettyStatusMessage() const override;
     Calamares::JobResult exec() override;
 
-    static QString make_salt(int length);
+    static QString make_salt(size_t length);
 
 private:
     QString m_userName;
